chore(swingset): split test out to a separate file
We think GC under Node.js is a bit squirrly in the face of AVA's
parallelism, and moving the intermittently-failing test into its own
file might help

refs #5266
