+++ packages/SwingSet/test/vat-admin/terminate/test-terminate.js
@@ -9,14 +9,7 @@ import {
   loadSwingsetConfigFile,
   buildKernelBundles,
 } from '../../../src/index.js';
-
-function capdata(body, slots = []) {
-  return harden({ body, slots });
-}
-
-function capargs(args, slots = []) {
-  return capdata(JSON.stringify(args), slots);
-}
+import { capargs } from '../../util.js';
 
 test.before(async t => {
   const kernelBundles = await buildKernelBundles();
@@ -159,38 +152,6 @@ test.serial('dispatches to the dead do not harm kernel', async t => {
       'done: Error: arbitrary reason',
       'm: live 2 failed: Error: vat terminated',
     ]);
-  }
-});
-
-test.serial('replay does not resurrect dead vat', async t => {
-  const configPath = new URL('swingset-no-zombies.json', import.meta.url)
-    .pathname;
-  const config = await loadSwingsetConfigFile(configPath);
-
-  const hostStorage1 = provideHostStorage();
-  {
-    const c1 = await buildVatController(config, [], {
-      hostStorage: hostStorage1,
-      kernelBundles: t.context.data.kernelBundles,
-    });
-    await c1.run();
-    t.deepEqual(c1.kpResolution(c1.bootstrapResult), capargs('bootstrap done'));
-    // this comes from the dynamic vat...
-    t.deepEqual(c1.dump().log, [`w: I ate'nt dead`]);
-  }
-
-  const state1 = getAllState(hostStorage1);
-  const hostStorage2 = provideHostStorage();
-  // XXX TODO also copy transcripts
-  setAllState(hostStorage2, state1);
-  {
-    const c2 = await buildVatController(config, [], {
-      hostStorage: hostStorage2,
-      kernelBundles: t.context.data.kernelBundles,
-    });
-    await c2.run();
-    // ...which shouldn't run the second time through
-    t.deepEqual(c2.dump().log, []);
   }
 });
